função separaUltimoNumDaCrip funcionando

+++ src/pixel.cpp
@@ -9,14 +9,14 @@ using namespace std;
 
 void Pixel::setRed(char red){
 	this->red = red;
-	this->redVector.push_back((int)red);
+	this->redVector.push_back((unsigned int)red);
 }
 char Pixel::getRed(){
 	return red;
 }
 
 int Pixel::getRedVector(int i){
-	return redVector[i];
+	return (unsigned)redVector[i];
 
 }
 
@@ -24,28 +24,28 @@ int Pixel::getRedVector(int i){
 
 void Pixel::setGreen(char green){
 	this->green = green;
-	this->greenVector.push_back((int)green);
+	this->greenVector.push_back((unsigned int)green);
 }
 char Pixel::getGreen(){
 	return green;
 }
 
 int Pixel::getGreenVector(int i){
-	return greenVector[i];
+	return (unsigned)greenVector[i];
 }
 
 
 
 void Pixel::setBlue(char blue){
 	this->blue = blue;
-	this->blueVector.push_back((int)blue);
+	this->blueVector.push_back((unsigned int)blue);
 }
 char Pixel::getBlue(){
 	return blue;
 }
 
 int Pixel::getBlueVector(int i){
-	return blueVector[i];
+	return (unsigned)blueVector[i];
 }
